Update Main.java
The following changes add a loop so that after the client sends and receives a string from the server it can re-send more until the users chooses to terminate the conncection by typing "exit". Currently, it successfully exits and sends the first sting to the server with no issues. However, the server does not see the strings after the first responce so changes will need to be done there.