Update test.java
Lots of changes to the overall structure here. 

The server is in an infinate loop while running looking for new clients. When our client connects the server creates a thread that keeps connection with the client code (till the client chooses to disconnect). 

I belive this method will make the next steps for keeping track of the information for each client seperatly easyier.